done without stemming, but with low accuracy

+++ src/nyu/predictiveanalytics/Entity.java
@@ -4,14 +4,14 @@ public class Entity {
 	
 	int m_featureDimension;
 	//int[] m_features;
-	float[] m_features;
+	double[] m_features;
 	
 	
 	public Entity(int dimension){
 		m_featureDimension = dimension;
 	}
 	
-	public Entity(float[] entity_features){
+	public Entity(double[] entity_features){
 		m_features = entity_features;
 		m_featureDimension = entity_features.length;
 	}
